LogParser/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.9.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
cssselect==1.1.0
decorator==5.1.1
dicttoxml==1.7.4
docutils==0.18.1
frozenlist==1.3.0
future==0.18.2
geocoder==1.38.1
geoip2==4.5.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.2
ip2geotools==0.1.6
IP2Location==8.6.4
isort==5.10.1
jeepney==0.7.1
keyring==23.5.0
lazy-object-proxy==1.7.1
lxml==4.8.0
maxminddb==2.2.0
mccabe==0.6.1
multidict==6.0.2
outcome==1.1.0
packaging==21.3
pip-review==1.1.1
pkginfo==1.8.2
platformdirs==2.5.1
pycodestyle==2.8.0
pycountry==22.1.10
pycparser==2.21
Pygments==2.11.2
pylint==2.12.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyquery==1.4.3
PySocks==1.7.1
PyYAML==6.0
ratelim==0.1.6
readme-renderer==32.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
SecretStorage==3.3.1
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
tqdm==4.63.0
trio==0.20.0
trio-websocket==0.9.2
twine==3.8.0
typed-ast==1.5.2
typing==3.7.4.3
typing-extensions==4.1.1
urllib3==1.26.8
webencodings==0.5.1
wrapt==1.13.3
wsproto==1.1.0
yarl==1.7.2
zipp==3.7.0
